+++ plugins/clima.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+# NAME: Clima Local
+# DESC: Muestra el clima actual
+curl -s "wttr.in/?format=3"